pyproject.toml
[project]
name = "llm-knowledge-graph-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.8.0",
    "neo4j>=6.1.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "pyvis>=0.3.2",
]
